Enhanced files: src/channel.c; include/msg.h
Adds a new command to ircd called KNOCK <channel> which allows users
from outside the channel which do not match any bans to have the server
NOTICE the channel operators in the channel so they know the user wants
to JOIN the channel.

This patch also has an additional feature of sending a NOTICE to the
channel operators when a user is INVITEd to the channel when it is
invite-only (+i).

 Syntax:
  :<sender> KNOCK <channel>

 Possible numerics to return:
  ERR_NEEDMOREPARAMS    ERR_CANNOTSENDTOCHAN
  ERR_NOSUCHCHANNEL

-- David-R <rientjes@mail.whidbey.net>

committed by db@db.net