removed REJECT_IPHONE from /quote info was replaced with X line
send.c send channel wall to channel, which promptly gets routed
as a channel message instead of a channel wall. TS4 used @#
and only sent to TS4 capable servers. We could add a CHANW capability
to do this, but its also nice having backward compatibility for now...
Perhaps for 6.1
kludged a fix into send.c for now, will not work with VARARGS